% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lyrics.R
\name{get_lyrics}
\alias{get_lyrics}
\title{get_lyrics}
\usage{
get_lyrics(album, musica)
}
\arguments{
\item{album}{Nome do album que contém a música desejada}

\item{musica}{Nome da música desejada}
}
\value{
Retorna a letra da música em formato HTML
}
\description{
Obtêm a letra da música e do album especificado
}
\examples{
get_lyrics(album = "Roda Gigante", musica = "Roda-Gigante")

}
